+++ src/pulumi/core-dtap/analytics/datafactory/__init__.py
@@ -5,3 +5,6 @@ platform_outputs["analytics"]["datafactory"] = {"factories": {}}
 
 from . import integrated_integration_runtime
 from . import orchestration
+from . import orchestration_linked_services
+from . import orchestration_datasets
+from . import orchestration_pipelines